ohlc_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
requests
python-dotenv
apscheduler